+++ public/sign-out.php
@@ -1,0 +1,5 @@
+<?php
+require_once '../app/core/init.php';
+$user = new User();
+$user->logout();
+Redirect::to('index.php');